requirements.txt
paho-mqtt>=1.6.1,<2.0
requests>=2.31.0
pyyaml>=6.0.1
numpy>=1.24.0
scipy>=1.11.0
